Added "--includemain" argument to "seer evm generate"
If specified, this adds a `main()` method to the generated code which
executes the CLI.
